Split default pool into separate file
